requirements.txt
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
Django==2.2.6
django-cors-headers==3.1.1
django-filter==2.2.0
djangorestframework==3.10.3
fuzzywuzzy==0.17.0
google-api-core==1.14.3
google-auth==1.6.3
google-cloud-vision==0.39.0
googleapis-common-protos==1.6.0
grpcio==1.24.1
idna==2.8
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.17.3
Pillow==6.2.0
protobuf==3.10.0
psycopg2==2.8.3
pyasn1==0.4.7
pyasn1-modules==0.2.7
PyDocParser==1.1.1
pyparsing==2.4.2
pyspellchecker==0.5.2
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
